development upd: rebuilt ball movement logic from scratch
Ball movement logic was updated to be more realistic: after the ball was
hitten it has start speed and negative boost. As the result the ball's
speed is decreasing each moment after it has been hitten

mv: global variables, function declarations and user types to TheBall.h
