+++ TheBall/TheBall/TheBall.vcxproj
@@ -147,6 +147,7 @@
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="resource.h" />
+    <ClInclude Include="TheBall.h" />
   </ItemGroup>
   <ItemGroup>
     <ResourceCompile Include="TheBall.rc" />
